RecyclerView implementaion with two buttons added. Static tabs added.
ToDo:
- Edit and Add button for each item needs an onclicklistener.
Edit button should open an inputwindow of some kind, and communicate whatever description is inputed, so that when addbutton is pressed it will be added correctly to the orderlist.
One solution could be to use the setTag() and getTag() to communicate information between them.

- Tab system needs implementation.
Either we use fragments to switch between each "list". But the easier way would be to use simple
onclicklisteners for each tab and switch out the RecyclerAdapter (pointing to the correct menulist) for the RecyclerView list