RecyclerView implementaion with two buttons added. Static tabs added.
ToDo:
- Edit and Add button for each item needs an onclicklistener.
Edit button should open an inputwindow of some kind, and communicate whatever description is inputed, so that when addbutton is pressed it will be added correctly to the orderlist.
One solution could be to use the setTag() and getTag() to communicate information between them.

- Tab system needs implementation.
Either we use fragments to switch between each "list". But the easier way would be to use simple
onclicklisteners for each tab and switch out the RecyclerAdapter (pointing to the correct menulist) for the RecyclerView list

Implemented ready button for pushing orders into next status.
- OrderList only has one static list. Needs a list for every table.
- Create API for OrderList for adding/changing elements instead of accessing a "naked" arraylist member. This way we can add elements in a sorted fashion.

Fixed some design stuff / colors.
Also fixed edit button to add items when input is done, also the keyboard automatically closes after we're done.