Added ApiCommunicator class to handle all api communication. Added BackgroundApiFetcher class that constantly runs in the background to fetch updates from the api.
Added a incomplete function to update the MasterOrderList with the list we get from the api.
- Unsure how this should be handled. Either we synchronize all clients or we let the clients only care about the orders they are "interested in".
- Also how do we know which OrderItem has been updated, from the clients side?

Changed OrderMenuRecyclerAdapter to not use a listener to get absolute position on clicks. Also changed constructor to take in a OrderList and the orderCounter instead of using set functions.

ToDo: Kitchen side api needs to be implemented so we can test that everything works. For now only fetching the a la carte menu has been tested more or less.