Added API call for fetching tables from database. We now create our tableList from the tables we get from api. Big problems with OrderItemAdapter and KitchenMenuAdapter not reacting properly to item removals in some special scenarios. Added a tableNumber variable to OrderItemAdapter to be able to fetch a new reference to the correct Order from KitchenOrderList for when our reference is no longer working.
The itemList reference in OrderItemAdapter is acting very very odd when an order is "finished", and by that i mean that all orderitems have been cooked and sent to the api so we remove the order from the list. This causes the itemList to reference to an array that is "sort of" no longer there. So when new orders come in that belong to that same table we can no longer remove orderitems from that list via the readybutton. However it reacts correctly to incoming orderitems??. But this problem is solved by requesting a new reference to KitchenOrderList using our tableNumber.

A new problem arises from the KitchenMenuAdapter however. The KitchenMenuAdapter provides the tablenumber to the OrderItemAdapter. But when an order gets removed from the KitchenMenuAdapter, something weird happens during its next onBindViewHolder call. Causing the OrderItemAdapter to get the wrong tableNumber causing it to get a null reference itemList :(, help.