Found a work-around for the OrderItemAdapter problem. We use the OrderItemAdapter.ViewHolder to request a new reference because the ViewHolder actually contains the correct stuff, unlike the OrderItemAdapter for whatever reason.
Changed TableChoiceActivity Adapter to display the text in different colors depending on the state of the orderlist underlying it. So if the orderlist is empty for a table, then that table's text will be gray in the TableChoiceActivity. Also some other stuff.

Redesigned some stuff in the kitchen menu view.
- Removed ordernumber because we arent using it.
- Changed some font size and text color to improve readability.

Added some comments because i suck at commenting while writing code apparently.